feat: add user-defined name to memo

+++ store/db/sqlite/memo.go
@@ -10,11 +10,11 @@ import (
 )
 
 func (d *DB) CreateMemo(ctx context.Context, create *store.Memo) (*store.Memo, error) {
-	fields := []string{"`creator_id`", "`content`", "`visibility`"}
-	placeholder := []string{"?", "?", "?"}
-	args := []any{create.CreatorID, create.Content, create.Visibility}
+	fields := []string{"`resource_name`", "`creator_id`", "`content`", "`visibility`"}
+	placeholder := []string{"?", "?", "?", "?"}
+	args := []any{create.ResourceName, create.CreatorID, create.Content, create.Visibility}
 
-	stmt := "INSERT INTO memo (" + strings.Join(fields, ", ") + ") VALUES (" + strings.Join(placeholder, ", ") + ") RETURNING `id`, `created_ts`, `updated_ts`, `row_status`"
+	stmt := "INSERT INTO `memo` (" + strings.Join(fields, ", ") + ") VALUES (" + strings.Join(placeholder, ", ") + ") RETURNING `id`, `created_ts`, `updated_ts`, `row_status`"
 	if err := d.db.QueryRowContext(ctx, stmt, args...).Scan(
 		&create.ID,
 		&create.CreatedTs,
@@ -31,29 +31,32 @@ func (d *DB) ListMemos(ctx context.Context, find *store.FindMemo) ([]*store.Memo
 	where, args := []string{"1 = 1"}, []any{}
 
 	if v := find.ID; v != nil {
-		where, args = append(where, "memo.id = ?"), append(args, *v)
+		where, args = append(where, "`memo`.`id` = ?"), append(args, *v)
+	}
+	if v := find.ResourceName; v != nil {
+		where, args = append(where, "`memo`.`resource_name` = ?"), append(args, *v)
 	}
 	if v := find.CreatorID; v != nil {
-		where, args = append(where, "memo.creator_id = ?"), append(args, *v)
+		where, args = append(where, "`memo`.`creator_id` = ?"), append(args, *v)
 	}
 	if v := find.RowStatus; v != nil {
-		where, args = append(where, "memo.row_status = ?"), append(args, *v)
+		where, args = append(where, "`memo`.`row_status` = ?"), append(args, *v)
 	}
 	if v := find.CreatedTsBefore; v != nil {
-		where, args = append(where, "memo.created_ts < ?"), append(args, *v)
+		where, args = append(where, "`memo`.`created_ts` < ?"), append(args, *v)
 	}
 	if v := find.CreatedTsAfter; v != nil {
-		where, args = append(where, "memo.created_ts > ?"), append(args, *v)
+		where, args = append(where, "`memo`.`created_ts` > ?"), append(args, *v)
 	}
 	if v := find.UpdatedTsBefore; v != nil {
-		where, args = append(where, "memo.updated_ts < ?"), append(args, *v)
+		where, args = append(where, "`memo`.`updated_ts` < ?"), append(args, *v)
 	}
 	if v := find.UpdatedTsAfter; v != nil {
-		where, args = append(where, "memo.updated_ts > ?"), append(args, *v)
+		where, args = append(where, "`memo`.`updated_ts` > ?"), append(args, *v)
 	}
 	if v := find.ContentSearch; len(v) != 0 {
 		for _, s := range v {
-			where, args = append(where, "memo.content LIKE ?"), append(args, fmt.Sprintf("%%%s%%", s))
+			where, args = append(where, "`memo`.`content` LIKE ?"), append(args, fmt.Sprintf("%%%s%%", s))
 		}
 	}
 	if v := find.VisibilityList; len(v) != 0 {
@@ -62,43 +65,43 @@ func (d *DB) ListMemos(ctx context.Context, find *store.FindMemo) ([]*store.Memo
 			placeholder = append(placeholder, "?")
 			args = append(args, visibility.String())
 		}
-		where = append(where, fmt.Sprintf("memo.visibility in (%s)", strings.Join(placeholder, ",")))
+		where = append(where, fmt.Sprintf("`memo`.`visibility` IN (%s)", strings.Join(placeholder, ",")))
 	}
 	if find.ExcludeComments {
-		where = append(where, "parent_id IS NULL")
+		where = append(where, "`parent_id` IS NULL")
 	}
 
 	orders := []string{}
 	if find.OrderByPinned {
-		orders = append(orders, "pinned DESC")
+		orders = append(orders, "`pinned` DESC")
 	}
 	if find.OrderByUpdatedTs {
-		orders = append(orders, "updated_ts DESC")
+		orders = append(orders, "`updated_ts` DESC")
 	} else {
-		orders = append(orders, "created_ts DESC")
+		orders = append(orders, "`created_ts` DESC")
 	}
-	orders = append(orders, "id DESC")
+	orders = append(orders, "`id` DESC")
 
 	fields := []string{
-		`memo.id AS id`,
-		`memo.creator_id AS creator_id`,
-		`memo.created_ts AS created_ts`,
-		`memo.updated_ts AS updated_ts`,
-		`memo.row_status AS row_status`,
-		`memo.visibility AS visibility`,
-		`IFNULL(memo_organizer.pinned, 0) AS pinned`,
-		`memo_relation.related_memo_id AS parent_id`,
+		"`memo`.`id` AS `id`",
+		"`memo`.`resource_name` AS `resource_name`",
+		"`memo`.`creator_id` AS `creator_id`",
+		"`memo`.`created_ts` AS `created_ts`",
+		"`memo`.`updated_ts` AS `updated_ts`",
+		"`memo`.`row_status` AS `row_status`",
+		"`memo`.`visibility` AS `visibility`",
+		"IFNULL(`memo_organizer`.`pinned`, 0) AS `pinned`",
+		"`memo_relation`.`related_memo_id` AS `parent_id`",
 	}
 	if !find.ExcludeContent {
-		fields = append(fields, `memo.content AS content`)
+		fields = append(fields, "`memo`.`content` AS `content`")
 	}
 
-	query := `SELECT ` + strings.Join(fields, ", ") + `
-		FROM memo
-		LEFT JOIN memo_organizer ON memo.id = memo_organizer.memo_id AND memo.creator_id = memo_organizer.user_id
-		FULL JOIN memo_relation ON memo.id = memo_relation.memo_id AND memo_relation.type = "COMMENT"
-		WHERE ` + strings.Join(where, " AND ") + `
-		ORDER BY ` + strings.Join(orders, ", ")
+	query := "SELECT " + strings.Join(fields, ", ") + "FROM `memo` " +
+		"LEFT JOIN `memo_organizer` ON `memo`.`id` = `memo_organizer`.`memo_id` AND `memo`.`creator_id` = `memo_organizer`.`user_id` " +
+		"FULL JOIN `memo_relation` ON `memo`.`id` = `memo_relation`.`memo_id` AND `memo_relation`.`type` = \"COMMENT\"" + " " +
+		"WHERE " + strings.Join(where, " AND ") + " " +
+		"ORDER BY " + strings.Join(orders, ", ")
 	if find.Limit != nil {
 		query = fmt.Sprintf("%s LIMIT %d", query, *find.Limit)
 		if find.Offset != nil {
@@ -117,6 +120,7 @@ func (d *DB) ListMemos(ctx context.Context, find *store.FindMemo) ([]*store.Memo
 		var memo store.Memo
 		dests := []any{
 			&memo.ID,
+			&memo.ResourceName,
 			&memo.CreatorID,
 			&memo.CreatedTs,
 			&memo.UpdatedTs,
@@ -143,24 +147,27 @@ func (d *DB) ListMemos(ctx context.Context, find *store.FindMemo) ([]*store.Memo
 
 func (d *DB) UpdateMemo(ctx context.Context, update *store.UpdateMemo) error {
 	set, args := []string{}, []any{}
+	if v := update.ResourceName; v != nil {
+		set, args = append(set, "`resource_name` = ?"), append(args, *v)
+	}
 	if v := update.CreatedTs; v != nil {
-		set, args = append(set, "created_ts = ?"), append(args, *v)
+		set, args = append(set, "`created_ts` = ?"), append(args, *v)
 	}
 	if v := update.UpdatedTs; v != nil {
-		set, args = append(set, "updated_ts = ?"), append(args, *v)
+		set, args = append(set, "`updated_ts` = ?"), append(args, *v)
 	}
 	if v := update.RowStatus; v != nil {
-		set, args = append(set, "row_status = ?"), append(args, *v)
+		set, args = append(set, "`row_status` = ?"), append(args, *v)
 	}
 	if v := update.Content; v != nil {
-		set, args = append(set, "content = ?"), append(args, *v)
+		set, args = append(set, "`content` = ?"), append(args, *v)
 	}
 	if v := update.Visibility; v != nil {
-		set, args = append(set, "visibility = ?"), append(args, *v)
+		set, args = append(set, "`visibility` = ?"), append(args, *v)
 	}
 	args = append(args, update.ID)
 
-	stmt := `UPDATE memo SET ` + strings.Join(set, ", ") + ` WHERE id = ?`
+	stmt := "UPDATE `memo` SET " + strings.Join(set, ", ") + " WHERE `id` = ?"
 	if _, err := d.db.ExecContext(ctx, stmt, args...); err != nil {
 		return err
 	}
@@ -168,8 +175,8 @@ func (d *DB) UpdateMemo(ctx context.Context, update *store.UpdateMemo) error {
 }
 
 func (d *DB) DeleteMemo(ctx context.Context, delete *store.DeleteMemo) error {
-	where, args := []string{"id = ?"}, []any{delete.ID}
-	stmt := `DELETE FROM memo WHERE ` + strings.Join(where, " AND ")
+	where, args := []string{"`id` = ?"}, []any{delete.ID}
+	stmt := "DELETE FROM `memo` WHERE " + strings.Join(where, " AND ")
 	result, err := d.db.ExecContext(ctx, stmt, args...)
 	if err != nil {
 		return err
@@ -186,16 +193,7 @@ func (d *DB) DeleteMemo(ctx context.Context, delete *store.DeleteMemo) error {
 }
 
 func vacuumMemo(ctx context.Context, tx *sql.Tx) error {
-	stmt := `
-	DELETE FROM 
-		memo 
-	WHERE 
-		creator_id NOT IN (
-			SELECT 
-				id 
-			FROM 
-				user
-		)`
+	stmt := "DELETE FROM `memo` WHERE `creator_id` NOT IN (SELECT `id` FROM `user`)"
 	_, err := tx.ExecContext(ctx, stmt)
 	if err != nil {
 		return err
